requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.55.3
google-auth==2.38.0
google-genai==1.2.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
flask-sqlalchemy==3.1.1
